+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -1,6 +1,6 @@
 package hello.controllers;
 
-import hello.EarthquakeQueryService;
+import hello.services.EarthquakeQueryService;
 import hello.EqSearch;
 
 import hello.geojson.FeatureCollection;
